refactor(serviceQueue): expose pending/size as runes; implement IOperationQueue (Sub-A.2.1)
