+++ src/components/componentQueueList.ts
@@ -1,6 +1,6 @@
 import type { StagedOp, VirtualFileState } from "../types/typeOps";
 import type { OpKind } from "../types/typeOps";
-import type { OperationQueueService } from "../services/serviceQueue";
+import type { OperationQueueService } from "../services/serviceQueue.svelte";
 import { translate } from "../i18n/index";
 
 export interface QueueListCallbacks {
