+++ src/components/views/viewDiff.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { stringifyYaml } from "obsidian";
   import { translate } from "../../i18n/index";
-  import type { OperationQueueService } from "../../services/serviceQueue";
+  import type { OperationQueueService } from "../../services/serviceQueue.svelte";
   import {
     buildDiff,
     buildOperationDiff,
