+++ src/modals/modalQueueDetails.ts
@@ -4,7 +4,7 @@
  * `twinkling-pearl.md`). Scheduled for deletion in Fase 5. Do not extend.
  */
 import { Modal, Notice, Setting, type App } from 'obsidian';
-import type { OperationQueueService } from '../services/serviceQueue';
+import type { OperationQueueService } from '../services/serviceQueue.svelte';
 import { translate } from '../i18n/index';
 import {
   buildDiff,
